fix(iOS): expose testID from wrapping view component.
Before this change, `testID` was exposed from the native sub component `IOSCheckBoxNativeComponent`.

However, this component does not expose the accessibility info (`accessible`, `accessibilityRole`, etc.), which is exposed in the parent RN component `View`.

Thus, retrieving this information on tests by the `testID` is not possible unless accessing the parent view.